requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.6.0
magic-filter==1.0.9
multidict==6.0.4
python-dotenv==0.21.1
pytz==2023.3
typing-extensions==4.5.0
websocket-client==1.5.1
yarl==1.9.2
zipp==3.15.0
